+++ src/modules/garment-purchasing/purchase-order-external/view.js
@@ -8,6 +8,7 @@ export class View {
     hasEdit = false;
     hasDelete = false;
     hasUnpost = false;
+    hasView=true;
 
     constructor(router, service) {
         this.router = router;
@@ -19,85 +20,46 @@ export class View {
         var id = params.id;
         this.poExId = id;
         this.data = await this.service.getById(id);
-        this.kurs = await this.service.getKurs(this.data.currency.code, this.data.date);
-
-        var getUsedBudget = [];
-        var getPRById = [];
-        var listPR = this.data.items.map((item) => {
-            return item.prId.toString()
-        });
-        var listPrIds = listPR.filter(function (elem, index, self) {
-            return index == self.indexOf(elem);
-        })
-        listPrIds.map((id) => {
-            getPRById.push(this.service.getPRById(id, ["no", "items.refNo", "items.quantity", "items.budgetPrice", "items.product.code"]))
-        });
-
-        for (var item of this.data.items) {
-            getUsedBudget.push(this.service.getListUsedBudget(item.prNo, item.prRefNo, item.product.code, this.data.no))
+        var kurs = await this.service.getKurs(this.data.Currency.Code, new Date(this.data.OrderDate).toLocaleDateString());
+        this.kurs=kurs[0];
+        var isUsedSJ=false;
+        for(var item of this.data.Items){
+            if(item.DOQuantity>0){
+                isUsedSJ=true;break;
+            }
         }
 
-        var poIds = this.data.items.map(function (item) {
-            return item.poId;
-        });
-        poIds = poIds.filter(function (elem, index, self) {
-            return index == self.indexOf(elem);
-        })
-
-        var getStatusPo = [];
-        for (var poId of poIds) {
-            getStatusPo.push(this.service.getPoId(poId, ["status.value"]))
+        if(this.data.Currency){
+            this.selectedCurrency=this.data.Currency;
         }
-        return Promise.all(getStatusPo)
-            .then((listStatusPo) => {
-                return Promise.all(getPRById)
-                    .then((listPR) => {
-                        return Promise.all(getUsedBudget)
-                            .then((listUsedBudget) => {
-                                listUsedBudget = [].concat.apply([], listUsedBudget);
-                                for (var item of this.data.items) {
-                                    var pr = listPR.find((pr) => pr.no.toString() == item.prNo.toString());
-                                    var prItem = pr.items.find((prItem) => prItem.product.code.toString() === item.product.code.toString() && prItem.refNo === item.prRefNo)
 
-                                    var budgetUsed = 0;
-                                    if (listUsedBudget.length > 0) {
-                                        var prevAmount = listUsedBudget.find((budget) => budget.prNo == item.prNo && budget.refNo == item.refNo && budget.product == item.product.code);
-                                        if (prevAmount) {
-                                            budgetUsed = budgetUsed + prevAmount.totalAmount;
-                                        }
-                                    }
-                                    item.budgetUsed = budgetUsed;
-                                    item.totalBudget = prItem.quantity * prItem.budgetPrice;
-                                }
+        if(this.data.Supplier){
+            this.selectedSupplier=this.data.Supplier;
+            this.data.SupplierId=this.data.Supplier.Id;
+            this.data.Supplier.usevat=this.data.IsUseVat ;
+           
+            if(this.data.IsIncomeTax){
+                this.data.Supplier.usetax=true;
+            }
+            
+        }
 
-                                if (this.data.status.value === 0) {
-                                    isVoid = true;
-                                }
-                                if (listStatusPo.find(po => { return po.status.value > 3 }) != undefined) {
-                                    isArriving = true;
-                                }
-                                if (!this.data.isPosted) {
-                                    this.hasDelete = true;
-                                    this.hasEdit = true;
-                                }
-                                if (this.data.isPosted && !isVoid && !isArriving && !this.data.isClosed) {
-                                    this.hasUnpost = true;
-                                }
+        if(this.data.IncomeTax){
+            this.selectedIncomeTax=this.data.IncomeTax;
+        }
 
-                                if (this.data.supplier) {
-                                    this.selectedSupplier = this.data.supplier;
-                                }
-                                if (this.data.currency) {
-                                    this.selectedCurrency = this.data.currency;
-                                }
-                                if (this.data.vat) {
-                                    this.selectedVat = this.data.vat;
-                                }
+        if (!this.data.IsPosted && !isUsedSJ) {
+            this.hasDelete = true;
+            this.hasEdit = true;
+        }
+        if (this.data.IsPosted && !isUsedSJ) {
+            this.hasUnpost = true;
+        }
+        if (this.data.IsCanceled || this.data.IsClosed) {
+            this.hasUnpost = false;
+        }
 
-                                return this.data;
-                            })
-                    })
-            })
+       
     }
 
     cancel(event) {
@@ -105,13 +67,17 @@ export class View {
     }
 
     edit(event) {
-        this.router.navigateToRoute('edit', { id: this.data._id });
+        this.router.navigateToRoute('edit', { id: this.data.Id });
     }
 
     delete(event) {
-        this.service.delete(this.data).then(result => {
-            this.cancel();
-        });
+        var r = confirm("Apakah Anda yakin akan menghapus data ini?");
+        if (r == true) {
+            this.service.delete(this.data).then(result => {
+                this.cancel();
+            });
+        } 
+        
     }
 
     cancelPO(e) {
